+++ legacy/bash/concat-clips.sh
@@ -1,1 +1,0 @@
-eval ffmpeg -i 0.webm -i 1.webm -filter_complex "gltransition=duration=4:offset=1.5" output.webm